arar arqqqpomlkjihgfedc_\ZYXWVUTSaSSRQPONMLKJIJaJKLMMNOPQRSTUU(V(W(X(Y(\())_))a)b)c)d)e*f*g*j*k*l*m+n+o+p+q+r+s+t+u+v+x,y,z,,,,,,,----------------- ---,,+**)(%((a(a(a((%a"a%()*++,---.. ......----.......r.s.s-r-q-p-p,o,n,m,l,k,j,i,h,f+e+d+c+b+a+*_***\***_**a*b*c*d*e*f*g*h*i*j*j+k+j+i+j+k+l+m+m,n,o,p,q,r,s-t-u-v-w-y-y.z................v.u.t.s.r.q.p.o.n.m.l.k.j.i.h.g.f.e.d.c.b.a.._...\..Z.X.W.V.U.T.S.R.Q.P.O.N.M.L.K.J.I.H.G.F.E.D.H.I.J.K.L.M.N.O.P.Q.S.T.U.V.W.X.Y.Z..\..._..a.b.c.d.e.f.g.h.j.k.l.m.n-o-p-q-r-s-u-v-w-x-y-z-z,,,,,,,,,,,,,,++++++++++++++++++++++++++++++++++++++++++++++++++++++++++*************))))))))))) ))))))))))))))))))))))))))))))))))))))))*********++++++++++++++++++++++++++++++++++++++*****z*y*x*w*v*u*pu*pupupu!quu!qu!uquuauu*au*au*u+t+s+r+q+p+o+n+m+l+k+j+i+h+h,g,f,e,d,c,b,a,,_,,,\,,Z,Y,X,W,V,U,T,T-S-R-Q-P-O.P.P-Q-Q.R.R-R,S,S+S*S)S(R(Q(Q)P)O)N)M)L)K)J)I)H)G)F)E)D)C)B)A))(?((?((A(B(C(D(E(F(G(H(I(J(K(L(M(N(O(P(Q(R(S(T(U(V(W(X(Y(Z((\((()_))a)b)c)d)e)f)g)h)i)j)k)l)m)n)o)p)q)r)s)t)u)v)w)x)y)z))))(((((((((%%%%%%%%%%%%%%%"!!!!!!!!!!!!!!!""""""""""""""""""""""""""""""""""""""""?"?"?"?"?"?"?"?"?"?"?"?"?"?"?"""""""""%%%%%%%%%%%%%%%%%%%%%(((((()))))))))))))))))))*********+++,,,,,-z-y-x-x.y.y-z--------,,,,,,,,,+++++++********)))))))))))))))(((((((((((((((((((((((((((((((((((((((((((((((((((((%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%z%y%x%w%v%u%t%s%r%q%p%o%n%m%l%k%j%i%h%g%f%e%d%c%d%e%f%g%h%i%j%k%l%m%n%o%p%q%r%s%t%u%v%w%x%y%z%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%((())**++,,,--...---,,++**))(%%% ORG-17: Implemented viewing sessions list, session data and another user's profiles.
"""""""%%%((((((((((((((((())))))))))************)))))))

+++ client/app/src/main/java/vsu/tp53/onboardapplication/home/PlayerModel.kt
@@ -1,13 +1,13 @@
 package vsu.tp53.onboardapplication.home
 
-class PlayerModel(_name: String, _reputation: Int) {
-    private var reputation: Int = _reputation
+class PlayerModel(_name: String, _reputation: Double) {
+    private var reputation: Double = _reputation
     private var name: String = _name
-    fun getReputation(): Int {
+    fun getReputation(): Double {
         return reputation
     }
 
-    fun setReputation(reputation: Int) {
+    fun setReputation(reputation: Double) {
         this.reputation = reputation
     }
